# api/openapi/v1.json
{
    "components": {
        "schemas": {
...
            "github_com_denisakp_ogoune_internal_dto_v1.CreateAnnouncementRequest": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "dismissible": {
                        "type": "boolean"
                    },
                    "severity": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorDetail"
                    }
                },
                "type": "object"
            },
...
            "github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_AnnouncementResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.AnnouncementResponse"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MetaResponse"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Ogoune uptime monitoring — Public REST API v1.",
        "title": "Ogoune Public API",
        "version": "1.0"
    },
...
    "paths": {
        "/announcements": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List active announcement banners",
                "tags": [
                    "announcements"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.CreateAnnouncementRequest",
                                        "summary": "body",
                                        "description": "Announcement"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Announcement",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_AnnouncementResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Publish an announcement banner",
                "tags": [
                    "announcements"
                ]
            }
        },
        "/announcements/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Announcement ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Retract an announcement banner",
                "tags": [
                    "announcements"
                ]
            }
        },
...
        "/dashboards/{id}/layout": {
            "put": {
                "parameters": [
                    {
                        "description": "Dashboard ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SaveLayoutRequest",
                                        "summary": "body",
                                        "description": "Ordered widgets"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Ordered widgets",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_DashboardResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Save a dashboard's widget layout (owner only)",
                "tags": [
                    "dashboards"
                ]
            }
        },
...
        "/integrations/alert-rules": {
            "get": {
                "parameters": [
                    {
                        "description": "Uptime alert threshold percent (default 99)",
                        "in": "query",
                        "name": "uptimeThreshold",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/yaml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Prometheus rules YAML"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Download config-derived Prometheus alerting rules (YAML)",
                "tags": [
                    "integrations"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:9596/api/v1"
        }
    ]
}